[ExportVerilog] Conservatively don't spill with disallowLocalVariables.
We want to spill expressions to temporaries if they are larger than
the maximum number of tokens per expression threshold. But, if
disallowLocalVariables is set, and we are in a procedural region, we
cannot just spill to a variable.

There doesn't seem to be an efficient way to see if the user of the
expression is inside a procedural region, so this conservatively does
not spill any time disallowLocalVariables is set. This can certainly
be improved to allow spilling in some cases.

In the general case, we want to be able to spill even when we are in a
procedural region. This will require us to be able to emit a wire in
the outer scope.