[PrepareForEmission] Add max expression size before creating a wire.
If an expression has more than the maximum number of terms, and it is
being hoisted out of a procedural region because local variables are
disallowed, always generate a wire. The intent is to remove the burden
of spilling expressions larger than the maximum token limit of some
simulators by guessing what's "too large" and doing it before
ExportVerilog. Better heuristics for when to create a wire can be
added.

The current implementation of prepareHWModule makes the phase ordering
of this a little interesting: we need to know it is "too large" before
we break the expression up into a binary tree, and we need to apply
this knowledge in a later iteration of the loop when it is actually
hoisted. For this, I maintained a set of Operation pointers.